+++ Adapter/ShopwareAdapter/ServiceBus/CommandHandler/Manufacturer/HandleManufacturerCommandHandler.php
@@ -2,10 +2,13 @@
 
 namespace ShopwareAdapter\ServiceBus\CommandHandler\Manufacturer;
 
+use Doctrine\ORM\EntityManagerInterface;
 use Shopware\Components\Api\Exception\NotFoundException as ManufacturerNotFoundException;
 use Shopware\Components\Api\Exception\ParameterMissingException;
 use Shopware\Components\Api\Exception\ValidationException;
 use Shopware\Components\Api\Resource\Manufacturer as ManufacturerResource;
+use Shopware\Models\Country\Country as CountryModel;
+use Shopware\Models\Country\Repository as CountryRepository;
 use ShopwareAdapter\DataPersister\Attribute\AttributeDataPersisterInterface;
 use ShopwareAdapter\ShopwareAdapter;
 use SystemConnector\IdentityService\IdentityServiceInterface;
@@ -13,8 +16,10 @@ use SystemConnector\ServiceBus\Command\CommandInterface;
 use SystemConnector\ServiceBus\Command\TransferObjectCommand;
 use SystemConnector\ServiceBus\CommandHandler\CommandHandlerInterface;
 use SystemConnector\ServiceBus\CommandType;
+use SystemConnector\TransferObject\Country\Country;
 use SystemConnector\TransferObject\Manufacturer\Manufacturer;
 use SystemConnector\TransferObject\Media\Media;
+use SystemConnector\ValueObject\Attribute\Attribute;
 
 class HandleManufacturerCommandHandler implements CommandHandlerInterface
 {
@@ -33,14 +38,27 @@ class HandleManufacturerCommandHandler implements CommandHandlerInterface
      */
     private $attributePersister;
 
+    /**
+     * @var EntityManagerInterface
+     */
+    private $entityManager;
+
+    /**
+     * @var CountryRepository
+     */
+    private $countryRepository;
+
     public function __construct(
         ManufacturerResource $resource,
         IdentityServiceInterface $identityService,
-        AttributeDataPersisterInterface $attributePersister
+        AttributeDataPersisterInterface $attributePersister,
+        EntityManagerInterface $entityManager
     ) {
         $this->resource = $resource;
         $this->identityService = $identityService;
         $this->attributePersister = $attributePersister;
+        $this->entityManager = $entityManager;
+        $this->countryRepository = $entityManager->getRepository(CountryModel::class);
     }
 
     /**
@@ -79,6 +97,23 @@ class HandleManufacturerCommandHandler implements CommandHandlerInterface
         if (null !== $manufacturer->getLink()) {
             $params['link'] = $manufacturer->getLink();
         }
+
+        $attributes = $manufacturer->getAttributes();
+
+        /** @var null|Attribute $countryIdentifierAttribute */
+        $countryIdentifierAttribute = array_filter($manufacturer->getAttributes(), function (Attribute $attribute) {
+            return $attribute->getKey() === 'countryIdentifier';
+        })[8] ?? null;
+
+        if (null !== $countryIdentifierAttribute) {
+            $countryNameAttribute = $this->getCountryNameAsAttribute($countryIdentifierAttribute->getValue());
+
+            if (null !== $countryNameAttribute) {
+                $attributes[] = $countryNameAttribute;
+            }
+        }
+
+        $manufacturer->setAttributes($attributes);
 
         if (null !== $manufacturer->getLogoIdentifier()) {
             $mediaIdentity = $this->identityService->findOneBy([
@@ -160,5 +195,35 @@ class HandleManufacturerCommandHandler implements CommandHandlerInterface
         }
 
         return array_shift($result['data']);
+    }
+
+    /**
+     * @param string $objectIdentifier
+     *
+     * @return null|Attribute
+     */
+    private function getCountryNameAsAttribute(string $objectIdentifier): ?Attribute
+    {
+        $countryIdentity = $this->identityService->findOneBy([
+            'objectIdentifier' => $objectIdentifier,
+            'objectType' => Country::TYPE,
+            'adapterName' => ShopwareAdapter::NAME,
+        ]);
+
+        if (null === $countryIdentity) {
+            return null;
+        }
+
+        /**
+         * @var Country $country
+         */
+        $country = $this->countryRepository->find($countryIdentity->getAdapterIdentifier());
+
+        $countryNameAttribute = new Attribute();
+        $countryNameAttribute->setKey('countryName');
+        $countryNameAttribute->setValue($country->getName());
+        $countryNameAttribute->setType('text');
+
+        return $countryNameAttribute;
     }
 }